+++ apps/web/src/components/forms/index.ts
@@ -1,5 +1,3 @@
-export type { PollDetailsData } from "./poll-details-form";
 export { PollDetailsForm } from "./poll-details-form";
-export type { PollOptionsData } from "./poll-options-form/poll-options-form";
 export { default as PollOptionsForm } from "./poll-options-form/poll-options-form";
 export * from "./types";
