+++ apps/web/src/components/poll/mobile-poll.tsx
@@ -24,9 +24,8 @@ import { YouAvatar } from "@/components/poll/you-avatar";
 import { useOptions, usePoll } from "@/components/poll-context";
 import { usePermissions } from "@/contexts/permissions";
 import { Trans, useTranslation } from "@/i18n/client";
-
-import { useVisibleParticipants } from "../participants-provider";
 import { useUser } from "../user-provider";
+import { useVisibleParticipants } from "../visibility";
 import GroupedOptions from "./mobile-poll/grouped-options";
 
 if (typeof window !== "undefined") {
