+++ apps/web/src/lib/datetime/relative-time.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useDateTimeConfig } from "@/lib/datetime/client";
-import type { DateInput } from "@/lib/datetime/format";
 import { formatRelativeTime } from "@/lib/datetime/format";
+import type { DateInput } from "@/lib/datetime/types";
 import { useHydrated } from "@/lib/datetime/use-hydrated";
 import { toISO } from "./utils";
