+++ apps/web/src/lib/datetime/schema.ts
@@ -1,11 +1,11 @@
 import * as z from "zod";
+import type { TimeFormat } from "./types";
 import { normalizeTimeZone } from "./utils";
 
-export const timeFormatSchema = z.enum(["hours12", "hours24"]);
-
-// The formatting layer's own TimeFormat; same literals as the Prisma enum so
-// user records assign directly, without coupling this layer to the database.
-export type TimeFormat = z.infer<typeof timeFormatSchema>;
+export const timeFormatSchema = z.enum([
+  "hours12",
+  "hours24",
+]) satisfies z.ZodType<TimeFormat>;
 
 export const timeZoneSchema = z
   .string()
